cephfs-shell: support cephfs-shell.conf
cephfs-shell options should reside in cephfs-shell.conf and not in
ceph.conf. Please note, unlike before, -c option of cephfs-shell must
take path to cephfs-shell.conf instead of path to ceph.conf.

This commit also updates the tests for cephfs-shell accordingly and
renames the variable config_path to shell_conf_path so that it's easy
to distinguish.

Signed-off-by: Rishabh Dave <ridave@redhat.com>