qa/vstart_runner.py: pass socket to mount command
Although passing socket is redundant, it keeps the code on the safer
side. Also, it's better if mounting via vstart_runner.py is independent
of the mount helper feature of getting monitor socket automatically
since latest mount helper should actually be present in one of the
standard locations in order to be used.

Also, rewrite the mount method of LocalKernelMount a bit to make it
readable.

Signed-off-by: Rishabh Dave <ridave@redhat.com>