qa/cephfs: allow not aborting when mount command fails
All mount methods accept a new argument named check_status that controls
whether the code aborts execution when the mount command fails. When the
code doesn't abort, it returns a tuple containing the exception that was
raised and the stdout and stderr of the shell executing the command. This
allows running a negative test where mount method is expected to fail.

Signed-off-by: Rishabh Dave <ridave@redhat.com>